move args definitions to new file
